# app/src/Controller/Profile/PhotoController.php
<?php

declare(strict_types=1);

namespace App\Controller\Profile;

use App\Controller\AuthAwareController;
use App\Request\Profile\UpdatePhotoRequest;
use App\Service\PhotoStorageService;
use App\Service\UserService;
use Psr\Http\Message\ResponseInterface;
use Psr\Log\LoggerInterface;
use Spiral\Auth\AuthContextInterface;
use Spiral\Http\ResponseWrapper;
use Spiral\Router\Annotation\Route;
use Spiral\Translator\Traits\TranslatorTrait;

final class PhotoController extends AuthAwareController
{
    use TranslatorTrait;

    public function __construct(
        AuthContextInterface $auth,
        protected readonly LoggerInterface $logger,
        protected readonly UserService $userService,
        protected readonly ResponseWrapper $response,
        protected readonly PhotoStorageService $photoStorageService,
    ) {
        parent::__construct($auth);
    }

    #[Route(route: '/profile/photo', name: 'profile.update.photo', methods: 'PUT', group: 'auth')]
    public function updatePhoto(UpdatePhotoRequest $request): ResponseInterface
    {
        $file = $request->getPhoto();

        $oldFileName = $this->user->photo;

        $fileName = $this->photoStorageService->storeUploadedProfilePhoto($file);

        if ($fileName === null) {
            return $this->response->json([
                'message' => $this->say('profile_photo_update_empty'),
            ], 500);
        }

        $this->user->photo = $fileName;

        try {
            $this->userService->store($this->user);
        } catch (\Throwable $exception) {
            $this->logger->error('Unable to store user', [
                'action' => 'profile.update.photo',
                'userId' => $this->user->id,
                'exception' => $exception,
            ]);

            return $this->response->json([
                'message' => $this->say('profile_photo_update_exception'),
                'error'   => $exception->getMessage(),
            ], 500);
        }

        if ($oldFileName !== null) {
            $this->photoStorageService->removeProfilePhoto($oldFileName);
        }

        return $this->response->json([
            'message' => $this->say('profile_photo_update_ok'),
            'fileName' => $fileName,
            'url' => $this->photoStorageService->getProfilePhotoPublicUrl($fileName),
        ], 200);
    }
}