# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Cash-Track API
  version: 1.0.0
  license:
    name: Proprietary
    url: https://cash-track.app
  description: |
    Cash-Track is a personal finance management SaaS. This API powers the mobile and web clients.

    ## Authentication
    Most endpoints require a **JWT Bearer token** obtained via `/v1/auth/login`, `/v1/auth/register`,
    or one of the social/passkey login flows.

    Tokens come in pairs:
    - **Access token** — short-lived (~1 hour). Send in `Authorization: Bearer <token>`.
    - **Refresh token** — long-lived (~7 days). Use `POST /v1/auth/refresh` to get a new pair.

    ## Email Confirmation
    After registration the user receives a confirmation email. Creating wallets, charges, and tags
    requires a confirmed email address. Use the `/v1/auth/email/confirmation` group to check status
    or resend the link.

    ## Error Responses
    Standard errors return an `Error` object. Validation failures return a `ValidationError`
    object with per-field messages (HTTP 422).

    ## Idempotency
    Most state-changing endpoints (`POST`/`PUT`/`PATCH`/`DELETE`) accept an optional
    `Idempotency-Key` request header — a client-generated UUIDv4. Retrying the exact same
    request with the same key within 24 hours replays the original response instead of
    re-executing it (`Idempotency-Replayed: true` is added to the replayed response). Reusing
...
    rejected (`400`). See `IdempotencyKeyParam`, `IdempotencyKeyConflict`, and
...

...

...

...
servers:
...
paths:
...
  /v1/auth/password/forgot:
    post:
      operationId: passwordForgotCreate
      summary: Request a password reset
      description: |
        Sends a password reset code to the provided email address if an account exists.
        Returns the same 200 body for unknown emails and for repeat requests inside the resend
        window (no mail is sent), so the response never reveals whether an account exists.
        422 is returned only for a missing or malformed email.
      tags: [Password]
      security: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKeyParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "200":
          headers:
            Idempotency-Replayed:
              $ref: "#/components/headers/IdempotencyReplayed"
          description: Request accepted (mail sent only if the account exists and is not throttled)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
        "422":
          $ref: "#/components/responses/ValidationOrIdempotencyMismatch"
        "409":
          $ref: "#/components/responses/IdempotencyKeyConflict"
  /v1/auth/password/reset:
    post:
      operationId: passwordReset
      summary: Reset password with code
      description: |
        Completes the password reset flow. The `code` is included in the reset email.
        On success the old password is replaced; existing sessions remain valid until
        their access tokens expire.
      tags: [Password]
      security: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKeyParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetRequest"
      responses:
        "200":
          headers:
            Idempotency-Replayed:
              $ref: "#/components/headers/IdempotencyReplayed"
          description: Password changed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
        "422":
          $ref: "#/components/responses/ValidationOrIdempotencyMismatch"
        "409":
          $ref: "#/components/responses/IdempotencyKeyConflict"
  # =========================================================================
  # PROFILE
  # =========================================================================
...
  /v1/profile/photo:
    put:
      operationId: profileUpdatePhoto
      summary: Upload profile photo
      description: |
        Replaces the user's profile photo. Accepts a JPEG/PNG/WebP image.
        Constraints: max 5 MB, dimensions between 50×50 and 5000×5000 px.
        The response includes the public CDN URL of the stored photo.
      tags: [Profile]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKeyParam"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [photo]
              properties:
                photo:
                  type: string
                  format: binary
                  description: Image file (JPEG / PNG / WebP, max 5 MB, 50–5000 px)
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "200":
          headers:
            Idempotency-Replayed:
              $ref: "#/components/headers/IdempotencyReplayed"
          description: Photo uploaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  fileName:
                    type: [string, "null"]
                  url:
                    type: string
                    format: uri
        "422":
          $ref: "#/components/responses/ValidationOrIdempotencyMismatch"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/IdempotencyKeyConflict"
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token obtained from `/v1/auth/login`, `/v1/auth/register`, or any social
        login endpoint. Tokens expire after ~1 hour. Refresh via `POST /v1/auth/refresh`.

  # -------------------------------------------------------------------------
  # PARAMETERS
  # -------------------------------------------------------------------------
  parameters:
...
    IdempotencyKeyParam:
      name: Idempotency-Key
      in: header
      required: false
      description: |
        Optional. A canonical lowercase UUIDv4 supplied by the client. Sending the same key
        again with the same method, path, and request body within 24 hours replays the first
        response verbatim (see the `Idempotency-Replayed` response header) instead of
        re-executing the request. Sending it again with a *different* body returns `422`
        instead of running — see the `IdempotencyKeyMismatch` response. A malformed value
        (not a canonical UUIDv4) returns `400`. Omitting the header entirely runs the request
        normally, with no deduplication.
      schema:
        type: string
        format: uuid
        pattern: "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
        example: "1e6c2a6a-9e2e-4d3a-8f2e-6a2f1b8c9d10"

  # -------------------------------------------------------------------------
  # SCHEMAS
  # -------------------------------------------------------------------------
  schemas:
...
    PasswordResetRequest:
      type: object
      required:
        - code
        - password
        - passwordConfirmation
      properties:
        code:
          type: string
          description: Reset code sent to the user's email
        password:
          type: string
          minLength: 6
          format: password
        passwordConfirmation:
          type: string
          minLength: 6
          format: password
          description: Must match `password`
...
    MessageResponse:
      type: object
      description: Simple message response for success acknowledgements.
      properties:
        message:
          type: string
    Error:
      type: object
      description: Standard error envelope.
      required:
        - message
      properties:
        error:
          type: string
          description: Machine-readable error code (optional)
        message:
          type: string
          description: Human-readable error description
    ValidationError:
      type: object
      description: Field-level validation failures (HTTP 422).
      required:
        - errors
      properties:
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          description: Map of field names to arrays of error messages
          example:
            email:
              - "Must be a valid email address"
            password:
              - "Must be at least 6 characters"

  # -------------------------------------------------------------------------
  # REUSABLE HEADERS
  # -------------------------------------------------------------------------
  headers:
    ApiVersion:
      description: |
        Git tag of the running API image (e.g. `v1.4.2`). Present on every response; omitted
        entirely (rather than sent empty) when the image was built without a tag.
      schema:
        type: string
    ApiSha:
      description: |
        Git commit SHA of the running API image. Present on every response; omitted entirely
        (rather than sent empty) when the image was built without a commit SHA.
      schema:
        type: string
    TraceId:
      description: |
        The active OpenTelemetry trace ID. Present on every response; omitted entirely (rather
        than sent empty or all-zeros) when tracing is disabled (no valid active span context).
      schema:
        type: string
    RetryAfter:
      description: Seconds to wait before retrying.
      schema:
        type: integer
    IdempotencyReplayed:
      description: |
        `true` when this response is a verbatim replay of the first request that used this
        `Idempotency-Key`, rather than a fresh execution. Absent on a fresh response.
      schema:
        type: string
        enum: ["true"]

  # -------------------------------------------------------------------------
  # REUSABLE RESPONSES
  # -------------------------------------------------------------------------
  responses:
    BadRequest:
      description: |
        Bad request — invalid input or business rule violation. Also returned when an
        `Idempotency-Key` header is present but not a canonical lowercase UUIDv4.
      headers:
        X-Ct-Api-Version:
          $ref: "#/components/headers/ApiVersion"
        X-Ct-Api-Sha:
          $ref: "#/components/headers/ApiSha"
        X-Ct-Trace-Id:
          $ref: "#/components/headers/TraceId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: INVALID_CREDENTIALS
            message: The provided credentials are incorrect.
    Unauthorized:
      description: Authentication required or token expired
      headers:
        X-Ct-Api-Version:
          $ref: "#/components/headers/ApiVersion"
        X-Ct-Api-Sha:
          $ref: "#/components/headers/ApiSha"
        X-Ct-Trace-Id:
          $ref: "#/components/headers/TraceId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            message: Unauthenticated.
...
    IdempotencyKeyConflict:
      description: |
        A request with this `Idempotency-Key` (same scope, method, path, and request body) is
        still being processed. Retry after the given delay; do not send a different body under
        the same key while this is in flight.
      headers:
        X-Ct-Api-Version:
          $ref: "#/components/headers/ApiVersion"
        X-Ct-Api-Sha:
          $ref: "#/components/headers/ApiSha"
        X-Ct-Trace-Id:
          $ref: "#/components/headers/TraceId"
        Retry-After:
          $ref: "#/components/headers/RetryAfter"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            message: The original request with this Idempotency-Key is still processing. Please retry shortly.
...
    ValidationOrIdempotencyMismatch:
      description: |
        `422` on this endpoint has two distinct causes, both possible since it accepts an
        `Idempotency-Key`:
          - **Field-level validation failure** — `ValidationError`-shaped, carries an `errors`
            map keyed by field name.
          - **Idempotency key reused with a different body** — `Error`-shaped, a plain
            `message`, no `errors` key.

        Discriminate by the presence of `errors` in the body, not by anything else — both
        causes share this single `422` status code.
      headers:
        X-Ct-Api-Version:
          $ref: "#/components/headers/ApiVersion"
        X-Ct-Api-Sha:
          $ref: "#/components/headers/ApiSha"
        X-Ct-Trace-Id:
          $ref: "#/components/headers/TraceId"
      content:
        application/json:
          schema:
            oneOf:
              - $ref: "#/components/schemas/ValidationError"
              - $ref: "#/components/schemas/Error"
    InternalError:
      description: |
        Unexpected server-side failure (database, mail transport, etc.). Logged and reported;
        safe to retry. Not cached by `Idempotency-Key`.
      headers:
        X-Ct-Api-Version:
          $ref: "#/components/headers/ApiVersion"
        X-Ct-Api-Sha:
          $ref: "#/components/headers/ApiSha"
        X-Ct-Trace-Id:
          $ref: "#/components/headers/TraceId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: Transport exception
            message: Unable to reset your password.
...
security:
  - bearerAuth: []